my_news_portal/requirements.txt
asgiref
sqlparse
tzdata
Django
django-allauth
django-filter
requests
google-generativeai
python-dotenv
